requirements.txt
python-dotenv
playwright